requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
plotly>=5.18.0
pyyaml>=6.0
xlrd>=2.0.1
openpyxl>=3.1.0
numpy>=1.26.0
